+++ backend/AAiT-backend-group-5/go.mod
@@ -1,0 +1,5 @@
+module github.com/aait.backend.g5.main/backend
+
+go 1.22.5
+
+require go.mongodb.org/mongo-driver v1.16.1
